requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
contextvars==2.4
dill==0.3.2
Django==3.0.8
django-extensions==3.0.2
django-redis==4.12.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.30
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
install==1.3.3
multiprocess==0.70.10
mysqlclient==2.0.1
p-tqdm==1.3.3
pathos==0.2.6
Pillow==7.2.0
pox==0.2.8
ppft==1.6.6.2
pypinyin==0.39.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.47.0
urllib3==1.25.9
